requirements.txt
Django==1.3.1
PIL==1.1.7
South==0.7.3
boto==2.2.2
distribute==0.6.24
django-tastypie==0.9.11
ipython==0.12
mimeparse==0.1.3
psycopg2==2.4.4
python-dateutil==1.5
readline==6.2.2
wsgiref==0.1.2
